attach.resource {bigmemory.sri}	R Documentation

Create an Object that Makes Use of a Shared Resource

Description

This generic function should be used to attach to a system-wide shared resource on a computer. It is used by other packages.

Usage

attach.resource(obj, ...)

Arguments

obj
a descriptor object describing the location and type of a shared resource.

...
other information needed for object-specific attaching.

Value

An R object that makes use of a share resource.

Examples

# Create a concrete function.
setMethod(
 "attach.resource", 
 signature(obj = "NULL"),
 function (obj, ...) {
   NULL
 }
)

# Describe it
attach.resource(NULL)

[Package bigmemory.sri version 0.1.8 Index]