describe {bigmemory.sri}	R Documentation

Create a Description of a Shared Resource

Description

This generic function should be used for creating a description of a shared resource and is used in conjunction with attach.resource() to utilize shared resources from an R session.

Usage

describe(x)

Arguments

x
the object to create the descriptor for.

Value

A user-defined description of a shared resource.

See Also

[bigmemory.sri::attach.resource()]

Examples

# Create a concrete function.
setMethod(
 "describe", 
 signature(x = "list"),
 function (x) {
   print(x$value)
 }
)

# Create an object to describe.
tst = list(loc = "in_ram_not_shared", value = "hello world")

# Describe it
describe(tst)

[Package bigmemory.sri version 0.1.8 Index]